pull components into separate files
