+++ frontend/src/components/Profile/Profile.js
@@ -1,0 +1,12 @@
+import UserInfo from "./UserInfo";
+import SportsList from "./SportsList";
+
+function Profile() {
+  return (
+    <div className="profile-container">
+      <UserInfo />
+      <SportsList />
+    </div>
+  );
+}
+export default Profile;
